--- apps/auth-service/src/common/interfaces/active-user.interface.ts
@@ -1,9 +1,0 @@
-export enum UserRole {
-  User = 1,
-  Admin = 2,
-}
-
-export interface ActiveUser {
-  userId: string;
-  roleId: UserRole;
-}
