--- apps/auth-service/src/common/interfaces/prisma-error.interface.ts
@@ -1,17 +1,0 @@
-export interface PrismaDriverError {
-  // P2002 fields
-  driverAdapterError?: {
-    cause?: {
-      constraint?: {
-        fields: string[];
-      };
-    };
-  };
-  target?: string[];
-
-  // P2025 fields
-  modelName?: string;
-  model?: string;
-  operation?: string;
-  relation?: string;
-}
