+++ apps/auth-service/src/modules/audit-log/audit-log.module.ts
