+++ apps/auth-service/src/modules/audit-log/audit-log.service.ts
